% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv.predict.R
\name{predFit.gam}
\alias{predFit.gam}
\alias{predFit.scam}
\title{GAM prediction}
\usage{
\method{predFit}{gam}(object, newdata, type = c("link", "response"),
  interval = c("none", "confidence", "prediction"), level = 0.95, ...)

\method{predFit}{scam}(object, newdata, type = c("link", "response"),
  interval = c("none", "confidence", "prediction"), level = 0.95, ...)
}
\description{
GAM prediction

SCAM prediction
}
