\name{HACSim-package}
\alias{HACSim-package}
\alias{HAC.simaic}
\alias{HAC.simboot.bisect}
\alias{HAC.simboot.Newton}
\alias{HAC.simest}
\alias{HAC.simfit}
\alias{HAC.simmodels}
\alias{HAC.simplot}
\alias{HAC.simpost}
\alias{HAC.simrep}
\docType{package}
\title{
\packageTitle{HACSim}
}
\description{
HACSim employs a novel method of iteratively generating species haplotype accumulation curves in order to assess sampling completeness based on the approach of Phillips et al. (2015). Inverse estimation methods for univariate semiparametric fitting of haplotype accumulation curves via Gaussian generalized additive models (GAMs), Gaussian shape-constrained additve models (SCAMs) and Gaussian processes/kriging models for the calculation of specimen sample sizes are available, including bootstrap estimation of bias, standard error and confidence intervals. Posterior simulation for sample size estimation via a multivariate Normal distribution is also included. 
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{HACSim}
\packageIndices{HACSim}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{HACSim}

Maintainer: \packageMaintainer{HACSim}
}
\references{
Phillips, J.D., Gwiazdowski, R.A., Ashlock, D. and Hanner, R. (2015). An exploration of sufficient sampling effort to describe intraspecific DNA barcode haplotype diversity: examples from the ray-finned fishes (Chordata: Actinopterygii). DNA Barcodes, 3: 66-73.
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
